Refactor CrowdmapID authentication - Refs #39
* Autowire the UserDao bean in CrowdmapIDAuthenticationProvider
and not in CrowdmapIDClient
